requirements.txt
absl-py==0.11.0
albumentations==1.1.0
appdirs==1.4.4
argcomplete==1.12.3
argon2-cffi==21.1.0
astunparse==1.6.3
attrs==21.2.0
audioread==2.1.9
av @ file:///home/conda/feedstock_root/build_artifacts/av_1611413270203/work
backcall==0.2.0
bleach==4.1.0
cached-property==1.5.2
cachetools==4.2.1
calc-assoc-cuda==0.0.0
calc-pixel-feats-cuda==0.0.0
calc-spixel-feats-cuda==0.0.0
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
cycler==0.10.0
debugpy==1.5.1
decorator==4.4.2
defusedxml==0.7.1
docopt==0.6.2
efficientnet-pytorch==0.6.3
einops==0.3.2
entrypoints==0.3
et-xmlfile==1.0.1
facenet-pytorch==2.5.2
fastprogress==1.0.0
flatbuffers==1.12
ftfy==6.0.3
fvcore==0.1.5.post20210515
gast==0.3.3
google-auth==1.25.0
google-auth-oauthlib==0.4.2
google-pasta==0.2.0
graphviz==0.16
grpcio==1.32.0
h5py==2.10.0
idna==2.10
imageio==2.9.0
imgaug==0.4.0
importlib-metadata==3.4.0
init-spixel-feats-cuda==0.0.0
iopath==0.1.8
ipykernel==6.5.0
ipython==7.29.0
ipython-genutils==0.2.0
ipywidgets==7.6.5
jdcal==1.4.1
jedi==0.18.0
Jinja2==3.0.2
joblib==1.0.1
jsonpatch==1.32
jsonpointer==2.1
jsonschema==4.1.2
jupyter-client==7.0.6
jupyter-core==4.9.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.2
keras==2.8.0
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
librosa==0.8.1
linformer==0.2.1
llvmlite==0.36.0
Markdown==3.3.3
MarkupSafe==2.0.1
matplotlib==3.3.4
matplotlib-inline==0.1.3
mistune==0.8.4
mkl-fft==1.0.14
mkl-random==1.0.4
mkl-service==2.3.0
munch==2.5.0
nbclient==0.5.4
nbconvert==6.2.0
nbformat==5.1.3
nest-asyncio==1.5.1
networkx==2.5
notebook==6.4.5
numba==0.53.1
numpy==1.19.5
oauthlib==3.1.0
olefile==0.46
opencv-python==4.5.1.48
opencv-python-headless==4.5.3.56
openpyxl==3.0.6
opt-einsum==3.3.0
packaging==21.0
pandas==1.2.2
pandocfilters==1.5.0
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow @ file:///tmp/build/80754af9/pillow_1609786740831/work
pipreqs==0.4.10
pooch==1.4.0
portalocker==2.3.0
pretrainedmodels==0.7.4
prometheus-client==0.12.0
prompt-toolkit==3.0.21
protobuf==3.14.0
ptflops==0.6.5
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
Pygments==2.10.0
pyparsing==2.4.7
pyrsistent==0.18.0
python-dateutil==2.8.1
pytorch-warmup==0.0.4
pytvision==0.0.0
pytz==2021.1
PyWavelets==1.1.1
PyYAML==5.4.1
pyzmq==22.1.0
qudida==0.0.4
regex==2022.1.18
rel-to-abs-index-cuda==0.0.0
requests==2.25.1
requests-oauthlib==1.3.0
resampy==0.2.2
rsa==4.7
scikit-fmm==2021.2.2
scikit-image==0.18.1
scikit-learn==1.0.1
scipy==1.6.0
seaborn==0.11.1
segmentation-models-pytorch==0.2.0
Send2Trash==1.8.0
Shapely==1.7.1
simplejson==3.17.2
six @ file:///tmp/build/80754af9/six_1605205313296/work
sklearn==0.0
smear-cuda==0.0.0
SoundFile==0.10.3.post1
tabulate==0.8.9
tensorboard==2.4.1
tensorboard-logger==0.1.0
tensorboard-plugin-wit==1.8.0
tensorboardX==2.4.1
tensorflow==2.4.1
tensorflow-estimator==2.4.0
termcolor==1.1.0
terminado==0.12.1
testpath==0.5.0
thop==0.0.31.post2005241907
threadpoolctl==2.1.0
tifffile==2021.2.1
timer==0.1.2
timm==0.4.12
torch==1.7.1
torchaudio==0.7.0a0+a853dff
torchfile==0.1.0
torchstat==0.0.7
torchsummary==1.5.1
torchvision==0.8.2
tornado==6.1
tqdm==4.62.2
traitlets==5.1.1
ttach==0.0.3
typing-extensions @ file:///tmp/build/80754af9/typing_extensions_1611751222202/work
urllib3==1.26.3
visdom==0.1.8.9
vit-pytorch==0.20.8
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.1.0
Werkzeug==1.0.1
widgetsnbextension==3.5.2
wrapt==1.12.1
yacs==0.1.8
yarg==0.1.9
zipp==3.4.0
